refactor(proxy): split WebSocket logic into ws_proxy.py
Extract WebSocket proxy, record builder, and message reconstruction
from proxy.py into a dedicated claude_tap/ws_proxy.py module.

- proxy.py drops from ~788 to ~400 lines
- ws_proxy.py contains ~408 lines of WS-specific logic
- Avoids circular import by using a local import in proxy_handler
- Updates imports in forward_proxy.py and tests/test_ws_proxy.py
